+++ src/http/middlewares/middelware.jwt.ts
@@ -1,9 +1,9 @@
 import express from 'express';
 import * as jwt from 'jsonwebtoken';
-import {Token} from '../../models/types/Token';
-import {_handleResponse} from '../common/HandleResponse';
-import {TokenPayload} from '../../models/types/gen/tokenPayload';
-import { PermissionRoles } from '../../models/types/gen/permissions';
+import {Token} from '../../models/types/model.token';
+import {_handleResponse} from '../common/common.responseHandler';
+import {TokenPayload} from '../../models/types/gen/gen.token';
+import { PermissionRoles } from '../../models/types/gen/gen.permissions';
 
 // global computation exp time
 const _expMax = Math.floor(Date.now() / 1000) + 60 * 60 * 2;
