+++ src/models/types/gen/gen.token.ts
@@ -1,4 +1,4 @@
-import { PermissionRoles } from "./permissions";
+import { PermissionRoles } from "./gen.permissions";
 
 export interface TokenPayload {
   accountID: string;
