+++ src/models/types/model.account.ts
@@ -1,5 +1,5 @@
-import {Identificable} from './gen/identificable';
-import {Direction} from './gen/direction';
+import {Identificable} from './gen/gen.identificable';
+import {Direction} from './gen/gen.direction';
 
 export const _tableName = 'Accounts';
 export interface TinyAccount extends Identificable, Direction {
